category/games/2026/02/20260210-230500-vector-mario-bros-hammer-throw/pyproject.toml
[project]
name = "vector-mario-bros-hammer-throw"
version = "0.1.0"
description = "Master the trajectory to knock down koopas with precision hammer tosses."
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
